+++ changelog.md
@@ -1,9 +1,15 @@
 # CHANGELOG
+
+### V0.0.6
+[+] Added a little gallery to display your saved paintings
+[~] Made painting save upon pressing 'quit'
+[+] Added 'resume' button
+
 ### V0.0.5
 [+] Added button class, can display text or images
 [+] Added buttons for switching colours and tips
 [+] Added pause menu with quit button (esc to toggle this menu)
-[+] Added tip outline 
+[+] Added tip outline to preview drawing area
 [~] Made background colour vary based on active screen
 [-] Removed keybinds for switching colours and tips